refactor: move createSnsUser method to SnsUserService
