+++ src/main/java/com/example/backend/service/user/SnsUserService.java
@@ -1,7 +1,10 @@
 package com.example.backend.service.user;
 
+import com.example.backend.domain.User;
 import com.example.backend.domain.enumType.UserType;
+import com.example.backend.dto.login.SnsJoinRequestDTO;
 import com.example.backend.dto.user.SnsUserDTO;
+import org.springframework.transaction.annotation.Transactional;
 
 
 public interface SnsUserService {
@@ -9,6 +12,9 @@ public interface SnsUserService {
     String getUserEmail(UserType userType, String accessToken);
 
     SnsUserDTO getUserInfo(UserType userType, String accessToken);
+
+    @Transactional
+    User createSnsUser(UserType userType, SnsJoinRequestDTO joinDTO);
 
     //Test
 //    TestKakaoToken requestToken(String code) throws IOException;
