+++ src/main/java/com/example/backend/service/user/SnsUserServiceImpl.java
@@ -1,9 +1,12 @@
 package com.example.backend.service.user;
 
+import com.example.backend.domain.User;
 import com.example.backend.domain.enumType.UserType;
+import com.example.backend.dto.login.SnsJoinRequestDTO;
 import com.example.backend.dto.user.SnsUserDTO;
 import com.example.backend.global.exception.InvalidUserInputException;
 import com.example.backend.global.exception.InvalidUserInputExceptionType;
+import com.example.backend.repository.UserRepository;
 import com.example.backend.service.user.userInfo.KakaoUserInfo;
 import com.example.backend.service.user.userInfo.UserInfo;
 import com.example.backend.service.user.userInfo.UserInfoFactory;
@@ -17,6 +20,7 @@ import lombok.extern.slf4j.Slf4j;
 import org.springframework.beans.factory.annotation.Value;
 import org.springframework.http.*;
 import org.springframework.stereotype.Service;
+import org.springframework.transaction.annotation.Transactional;
 import org.springframework.util.LinkedMultiValueMap;
 import org.springframework.util.MultiValueMap;
 import org.springframework.web.client.RestTemplate;
@@ -31,6 +35,7 @@ public class SnsUserServiceImpl implements SnsUserService {
     private String KAKAO_URL;
 
     private final RestTemplate restTemplate;
+    private final UserRepository userRepository;
 
     @Override
     public String getUserEmail(UserType userType, String accessToken) {
@@ -70,6 +75,26 @@ public class SnsUserServiceImpl implements SnsUserService {
         }
 
         return UserInfoFactory.getUserDTOFromUserInfo(userType, userInfo);
+    }
+
+    @Transactional
+    @Override
+    public User createSnsUser(UserType userType, SnsJoinRequestDTO joinDTO) {
+        //사용자 정보 API에서 받아옴
+        SnsUserDTO userDTO = getUserInfo(userType, joinDTO.getAccessToken());
+        validateSocialUserDuplicate(userDTO.getEmail());
+        //중복X -> region 세팅 및 회원가입 처리
+        User user = userDTO.toEntity(
+                userType, joinDTO.getUsername(), joinDTO.getPhoneNumber(),
+                joinDTO.getDongCode(), joinDTO.getDongName());
+        user.setKakaoUser(); //status 세팅
+        return userRepository.save(user);
+    }
+
+    private void validateSocialUserDuplicate(String email){
+        if(userRepository.existsByEmail(email)){
+            throw new InvalidUserInputException(InvalidUserInputExceptionType.ALREADY_EXISTS_KAKAO_USER);
+        }
     }
 
 }
